requirements.txt
asgiref==3.8.1
Django==4.2
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.15.1
pillow==10.3.0
python-decouple==3.8
sqlparse==0.5.0
